Refactored EndpointController to make it pluggable.
